added database

+++ src/main/java/com/example/calculator_app/Calculator.java
@@ -1,23 +1,23 @@
 package com.example.calculator_app;
 public class Calculator{
-    public int add(int n1,int n2){
-        int result = n1+n2;
+    public Double add(Double n1,Double n2){
+        Double result = n1+n2;
         return result;
     }
-    public int subtract(int n1,int n2){
-        int result = n1-n2;
+    public Double subtract(Double n1,Double n2){
+        Double result = n1-n2;
         return result;
     }
-    public int multiply(int n1,int n2){
-        int result = n1*n2;
+    public Double multiply(Double n1,Double n2){
+        Double result = n1*n2;
         return result;
     }
-    public float divide(int n1,int n2){
-        float result = (float)n1/n2;
+    public Double divide(Double n1,Double n2){
+        Double result = (Double)n1/n2;
         return result;
     }
-    public int mod(int n1,int n2){
-        int result = n1%n2;
+    public Double mod(Double n1,Double n2){
+        Double result = n1%n2;
         return result;
     }
