+++ src/main/java/com/example/calculator_app/CalculatorRequest.java
@@ -1,18 +1,18 @@
 package com.example.calculator_app;
 public class CalculatorRequest{
-    private int num1;
-    private int num2;
+    private Double num1;
+    private Double num2;
     private String operation;
-    public void setNum1(int num1){
+    public void setNum1(Double num1){
         this.num1 = num1;
     }
-    public int getNum1(){
+    public Double getNum1(){
         return num1;
     }
-    public void setNum2(int num2){
+    public void setNum2(Double num2){
         this.num2 = num2;
     }
-    public int getNum2(){
+    public Double getNum2(){
         return num2;
     }
     public void setOperation(String operation){
